+++ lib/televerse.dart
@@ -24,6 +24,7 @@ part 'src/televerse/filters/filters.dart';
 part 'src/televerse/context/mixins/message_mixin.dart';
 part 'src/televerse/context/mixins/management.dart';
 part 'src/televerse/context/mixins/callback_mixin.dart';
+part 'src/televerse/context/mixins/inline_mixin.dart';
 
 /// The main class of the library.
 ///